+++ apps/meteor/client/views/marketplace/AppDetailsPage/tabs/AppLogs/AppLogs.tsx
@@ -1,6 +1,6 @@
+import type { ILogItem } from '@rocket.chat/core-typings';
 import { Box, Pagination } from '@rocket.chat/fuselage';
-import { useDebouncedValue } from '@rocket.chat/fuselage-hooks';
-import { useMemo, type ReactElement } from 'react';
+import { useEffect, useMemo, useReducer, type ReactElement } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import AppLogsItem from './AppLogsItem';
@@ -14,6 +14,25 @@ import { usePagination } from '../../../../../components/GenericTable/hooks/useP
 import AccordionLoading from '../../../components/AccordionLoading';
 import { useLogs } from '../../../hooks/useLogs';
 
+function expandedReducer(
+	expandedStates: { id: string; expanded: boolean }[],
+	action: { type: 'update'; id: string; expanded: boolean } | { type: 'expand-all' } | { type: 'reset'; logs: ILogItem[] },
+) {
+	switch (action.type) {
+		case 'update':
+			return expandedStates.map((state) => (state.id === action.id ? { ...state, expanded: action.expanded } : state));
+
+		case 'expand-all':
+			return expandedStates.map((state) => ({ ...state, expanded: true }));
+
+		case 'reset':
+			return action.logs.map((log) => ({ id: log._id, expanded: false }));
+
+		default:
+			return expandedStates;
+	}
+}
+
 const AppLogs = ({ id }: { id: string }): ReactElement => {
 	const { t } = useTranslation();
 
@@ -23,18 +42,28 @@ const AppLogs = ({ id }: { id: string }): ReactElement => {
 
 	const { current, itemsPerPage, setItemsPerPage: onSetItemsPerPage, setCurrent: onSetCurrent, ...paginationProps } = usePagination();
 
-	const debouncedEvent = useDebouncedValue(event, 500);
+	const [expandedStates, dispatch] = useReducer(expandedReducer, []);
 
-	const { data, isSuccess, isError, isFetching, error } = useLogs({
+	const handleExpand = ({ id, expanded }: { id: string; expanded: boolean }) => dispatch({ id, expanded, type: 'update' });
+
+	const handleExpandAll = () => dispatch({ type: 'expand-all' });
+
+	const { data, isSuccess, isError, error, refetch, isFetching } = useLogs({
 		appId: id,
 		current,
 		itemsPerPage,
 		...(instance !== 'all' && { instanceId: instance }),
 		...(severity !== 'all' && { logLevel: severity }),
-		method: debouncedEvent,
+		...(event !== 'all' && { method: event }),
 		...(startTime && startDate && { startDate: new Date(`${startDate}T${startTime}`).toISOString() }),
 		...(endTime && endDate && { endDate: new Date(`${endDate}T${endTime}`).toISOString() }),
 	});
+
+	useEffect(() => {
+		if (isSuccess) {
+			dispatch({ type: 'reset', logs: data.logs });
+		}
+	}, [data, isSuccess]);
 
 	const parsedError = useMemo(() => {
 		if (error) {
@@ -50,16 +79,29 @@ const AppLogs = ({ id }: { id: string }): ReactElement => {
 	return (
 		<>
 			<Box pb={16}>
-				<AppLogsFilter />
+				<AppLogsFilter
+					appId={id}
+					noResults={isFetching || !isSuccess || data?.logs?.length === 0}
+					isLoading={isFetching}
+					expandAll={() => handleExpandAll()}
+					refetchLogs={() => refetch()}
+				/>
 			</Box>
 			{isFetching && <AccordionLoading />}
 			{isError && <GenericError title={parsedError} />}
-			{isSuccess && data?.logs?.length === 0 ? (
-				<GenericNoResults />
-			) : (
+			{!isFetching && isSuccess && data?.logs?.length === 0 && <GenericNoResults />}
+			{!isFetching && isSuccess && data?.logs?.length > 0 && (
 				<CustomScrollbars>
-					<CollapsiblePanel aria-busy={isFetching || event !== debouncedEvent} width='100%' alignSelf='center'>
-						{data?.logs?.map((log, index) => <AppLogsItem regionId={log._id} key={`${index}-${log._createdAt}`} {...log} />)}
+					<CollapsiblePanel width='100%' alignSelf='center'>
+						{data?.logs?.map((log, index) => (
+							<AppLogsItem
+								regionId={log._id}
+								expanded={expandedStates.find((state) => state.id === log._id)?.expanded || false}
+								onExpand={handleExpand}
+								key={`${index}-${log._createdAt}`}
+								{...log}
+							/>
+						))}
 					</CollapsiblePanel>
 				</CustomScrollbars>
 			)}