+++ apps/meteor/client/views/marketplace/AppDetailsPage/tabs/AppLogs/useAppLogsFilterForm.ts
@@ -12,6 +12,6 @@ export type AppLogsFilterFormData = {
 };
 
 export const useAppLogsFilterForm = () =>
-	useForm<AppLogsFilterFormData>({ defaultValues: { severity: 'all', instance: 'all', timeFilter: 'all' } });
+	useForm<AppLogsFilterFormData>({ defaultValues: { severity: 'all', instance: 'all', timeFilter: 'all', event: 'all' } });
 
 export const useAppLogsFilterFormContext = () => useFormContext<AppLogsFilterFormData>();