+++ apps/meteor/client/views/navigation/sidepanel/SidePanelRouter.tsx
@@ -2,8 +2,11 @@ import SidePanelAll from './tabs/SidePanelAll';
 import { ALL_GROUPS, useRoomsListContext, useSidePanelFilter } from '../contexts/RoomsNavigationContext';
 import SidePanelDiscussions from './tabs/SidePanelDiscussions';
 import SidePanelFavorites from './tabs/SidePanelFavorites';
+import SidepanelInProgress from './tabs/SidePanelInProgress';
 import SidePanelMentions from './tabs/SidePanelMentions';
+import SidePanelQueue from './tabs/SidePanelQueue';
 import SidePanelRooms from './tabs/SidePanelRooms';
+import SidePanelOnHold from './tabs/SidepanelOnHold';
 
 const SidePanelRouter = () => {
 	const [currentTab] = useSidePanelFilter();
@@ -22,6 +25,12 @@ const SidePanelRouter = () => {
 		case ALL_GROUPS.CHANNELS:
 		case ALL_GROUPS.DIRECT_MESSAGES:
 			return parentRid ? <SidePanelRooms parentRid={parentRid} /> : null;
+		case ALL_GROUPS.IN_PROGRESS:
+			return <SidepanelInProgress />;
+		case ALL_GROUPS.ON_HOLD:
+			return <SidePanelOnHold />;
+		case ALL_GROUPS.QUEUE:
+			return <SidePanelQueue />;
 		default:
 			return null;
 	}