+++ apps/meteor/client/views/navigation/sidepanel/tabs/SidePanelInProgress.tsx
@@ -1,0 +1,28 @@
+import type { SubscriptionWithRoom } from '@rocket.chat/ui-contexts';
+import { useTranslation } from 'react-i18next';
+
+import {
+	SIDE_PANEL_GROUPS,
+	sidePanelFiltersConfig,
+	useSidePanelRoomsListTab,
+	useUnreadOnlyToggle,
+} from '../../contexts/RoomsNavigationContext';
+import SidePanel from '../SidePanel';
+
+const SidePanelInProgress = () => {
+	const { t } = useTranslation();
+	const rooms = useSidePanelRoomsListTab(SIDE_PANEL_GROUPS.IN_PROGRESS);
+	const [unreadOnly, toggleUnreadOnly] = useUnreadOnlyToggle();
+
+	return (
+		<SidePanel
+			title={t(sidePanelFiltersConfig[SIDE_PANEL_GROUPS.IN_PROGRESS].title)}
+			currentTab={SIDE_PANEL_GROUPS.IN_PROGRESS}
+			unreadOnly={unreadOnly}
+			toggleUnreadOnly={toggleUnreadOnly}
+			rooms={rooms as SubscriptionWithRoom[]}
+		/>
+	);
+};
+
+export default SidePanelInProgress;