+++ packages/apps-engine/src/definition/outboundComunication/IOutboundCommsProvider.ts
@@ -1,9 +1,10 @@
+import type { IHttp, IModify, IPersistence, IRead } from '../accessors';
 import type { IOutboundMessage } from './IOutboundMessage';
 import type { IOutboundProviderTemplate } from './IOutboundProviderTemplate';
 
-type ProviderMetadata = {
-	appId: string;
-	appName: string;
+export type ProviderMetadata = {
+	providerId: string;
+	providerName: string;
 	providerType: 'phone' | 'email';
 	supportsTemplates: boolean; // Indicates if the provider uses templates or not
 	templates: Record<string, IOutboundProviderTemplate[]>; // Format: { '+1121221212': [{ template }] }
@@ -14,12 +15,12 @@ interface IOutboundMessageProviderBase {
 	name: string;
 	documentationUrl?: string;
 	supportsTemplates?: boolean;
-	sendOutboundMessage(message: IOutboundMessage): Promise<void>;
+	sendOutboundMessage(message: IOutboundMessage, read: IRead, modify: IModify, http: IHttp, persistence: IPersistence): Promise<void>;
 }
 
 export interface IOutboundPhoneMessageProvider extends IOutboundMessageProviderBase {
 	type: 'phone';
-	getProviderMetadata(): Promise<ProviderMetadata>;
+	getProviderMetadata(read: IRead, modify: IModify, http: IHttp, persistence: IPersistence): Promise<ProviderMetadata>;
 }
 
 /*
@@ -30,3 +31,7 @@ export interface IOutboundEmailMessageProvider extends IOutboundMessageProviderB
 }
 
 export type IOutboundMessageProviders = IOutboundPhoneMessageProvider | IOutboundEmailMessageProvider;
+
+export const ValidOutboundProviderList = ['phone', 'email'] as const;
+
+export type ValidOutboundProvider = (typeof ValidOutboundProviderList)[number];