+++ packages/gazzodown/src/elements/Timestamp/index.tsx
@@ -1,18 +1,16 @@
-/* eslint-disable @typescript-eslint/explicit-function-return-type */
 import { Tag } from '@rocket.chat/fuselage';
 import type * as MessageParser from '@rocket.chat/message-parser';
-import { format } from 'date-fns';
-import { useContext, useEffect, useState, type ReactElement } from 'react';
+import { format, intlFormatDistance } from 'date-fns';
+import { useContext, useEffect, useState } from 'react';
 import { ErrorBoundary } from 'react-error-boundary';
 
-import { timeAgo } from './timeago';
 import { MarkupInteractionContext } from '../../MarkupInteractionContext';
 
 type BoldSpanProps = {
 	children: MessageParser.Timestamp;
 };
 
-const Timestamp = ({ format, value }: { format: 't' | 'T' | 'd' | 'D' | 'f' | 'F' | 'R'; value: Date }): ReactElement => {
+const Timestamp = ({ format, value }: { format: 't' | 'T' | 'd' | 'D' | 'f' | 'F' | 'R'; value: Date }) => {
 	switch (format) {
 		case 't': // Short time format
 			return <ShortTime value={value} />;
@@ -37,7 +35,7 @@ const Timestamp = ({ format, value }: { format: 't' | 'T' | 'd' | 'D' | 'f' | 'F
 };
 
 // eslint-disable-next-line react/no-multi-comp
-const TimestampWrapper = ({ children }: BoldSpanProps): ReactElement => {
+const TimestampWrapper = ({ children }: BoldSpanProps) => {
 	const { enableTimestamp } = useContext(MarkupInteractionContext);
 
 	if (!enableTimestamp) {
@@ -87,12 +85,12 @@ const RelativeTime = ({ value }: { value: Date }) => {
 	const time = value.getTime();
 
 	const { language } = useContext(MarkupInteractionContext);
-	const [text, setTime] = useState(() => timeAgo(time, language ?? 'en'));
+	const [text, setTime] = useState(() => intlFormatDistance(time, Date.now(), { locale: language ?? 'en' }));
 	const [timeToRefresh, setTimeToRefresh] = useState(() => getTimeToRefresh(time));
 
 	useEffect(() => {
 		const interval = setInterval(() => {
-			setTime(timeAgo(value.getTime(), 'en'));
+			setTime(intlFormatDistance(value.getTime(), Date.now(), { locale: 'en' }));
 			setTimeToRefresh(getTimeToRefresh(time));
 		}, timeToRefresh);