+++ packages/model-typings/src/models/ILivechatInquiryModel.ts
@@ -36,7 +36,7 @@ export interface ILivechatInquiryModel extends IBaseModel<ILivechatInquiryRecord
 	queueInquiry(inquiryId: string, lastMessage?: IMessage): Promise<ILivechatInquiryRecord | null>;
 	queueInquiryAndRemoveDefaultAgent(inquiryId: string): Promise<UpdateResult>;
 	readyInquiry(inquiryId: string): Promise<UpdateResult>;
-	changeDepartmentIdByRoomId(rid: string, department: string): Promise<void>;
+	changeDepartmentIdByRoomId(rid: string, department: string): Promise<UpdateResult>;
 	getStatus(inquiryId: string): Promise<ILivechatInquiryRecord['status'] | undefined>;
 	updateVisitorStatus(token: string, status: ILivechatInquiryRecord['v']['status']): Promise<UpdateResult>;
 	setDefaultAgentById(inquiryId: string, defaultAgent: ILivechatInquiryRecord['defaultAgent']): Promise<UpdateResult>;