+++ packages/api-client/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @rocket.chat/api-client
 
+## 0.2.31
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@7.8.2
+  - @rocket.chat/rest-typings@7.8.2
+  </details>
+
+## 0.2.30
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@7.8.1
+  - @rocket.chat/rest-typings@7.8.1
+  </details>
+
 ## 0.2.29
 
 ### Patch Changes